Intersection of two Linked Lists
Tails of two linked lists should be the same. First we cut the longer
list until it has same length with the shorter one. After that we start
moving both of them and comparing nodes if they are equal. At the end
if one of the nodes is null then there’s no intersection. Otherwise
return any of the current nodes.
Some sample tests are included.