Leetcode: 198. House Robber
Dynamic Programming solution: at each step we calculate 2 amounts - first if we want to rob this house, second if we don't. Obviously if we don't want to rob current house the max value would be maximum from previous two amounts. On the other hand, if we want to rob current house, then we should have not robbed the previous house, and add total cost of current house.

Eventually when we reach the end we just need to get the max of those two amounts.