Merge Sorted Arrays with Heaps
As an input you're given list of sorted arrays. The problem is to merge them efficiently. In this algorithm we're using Heaps, which gives us O(Nlog(K)) complexity, where K is the number of lists, and N is the total number of elements.

The approach is as following:
- we collect all iterators for each of the lists
- take first elements from each list and put them into min-heap
- then poll all elements from the heap one by one
- on every poll we get list id from where it was put initially
- move iterator on that list and new element to the queue

As a result we'll get sorted list.