2D Spiral Array problem
Given N return matrix NxN that prints number in a spiral way. Three different solutions are provided:
1) brute-force
2) "bump" method
3) recursive "bump" method

First approach is based on number of times we need to go each time. For example, if N=5, then the logic is as following:
- right 5 times
- down 4 times
- left 4 times
- up 3 times
- right 3 times
- down 2 times
- left 2 times
- up 1 time
- right 1 time

As we can see, there's a pattern:
n + (n-1)*2 + (n-2)*2 + ... + 1*2, which sums to n*n in total, so on every direction we need to make n - i steps 2 times. Only first iteration is n times, the rest are n-1 iterations are 2 times.

Second approach is based on "bumping" to dissalowed area. Possible scenarios for bumping:
- reached negative index
- reached overflow index
- reached already marked cell

Directions changing approach is same as from previous method.

The third solution is very similar to second one, with only difference of not running the while loop, but making a recursive call and passing indexes as params.